adding SweetAlert to button delete

+++ src/components/Product.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 //Redux
 import { useDispatch } from 'react-redux';
@@ -14,10 +15,20 @@ const Product = ({product}) => {
     const confirmRemoveProduct = id => {
 
         //ask the user
-
-
-        //pass it to action 
-        dispatch( removeProductAction(id) );
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.value) {
+                //pass it to action 
+                dispatch( removeProductAction(id) );  
+            }
+          });
     }
 
     return (
